\name{det_hand}
\alias{det_hand}
\title{Determine whether the left, right, or both arms were moved}
\usage{
det_hand(vel_x, vel_y)
}
\arguments{
\item{vel_x}{Data frame of x-axis velocity generated by velocity_x().}
\item{vel_y}{Data frame of y-axis velocity generated by velocity_y().}
}
\description{
Determine wheter the left, right, or both arms moved. The idea is two determine whether a sign/gesture is one or two-handed (in clips of single signs). May also be used to determine the hand that was primarily used in longer clips of discourse.

Computes Euclidean norms only for hand tracking points from BODY25 model and sees whether the velocity differs for both arms.
}
\examples{
det_hand(x, y)
}
