\name{en_acceleration}
\alias{en_acceleration}
\title{Compute Euclidean norm of sums of acceleration vectors (x,y)}
\usage{
en_acceleration(acc_x, acc_y)
}
\arguments{
\item{acc_x}{Data frame of x-axis acceleration generated by acceleration_x().}
\item{acc_y}{Data frame of y-axis acceleration generated by acceleration_y().}
\item{start_from_zero}{Determines if the output should be preceded with a row containing 0.}
}
\description{
Compute Euclidean norm of sums of acceleration vectors (x,y) for a tracked video using the x-axis and y-axis data generated using acceleration_x() and acceleration_y(). This is done using the following formula: sqrt(accelerationx(t+1) - accelerationx(t))^2 + (accelerationy(t+1) - accelerationy(t))^2)
}
\examples{
en_acceleration(x, y)
}
