\name{en_velocity}
\alias{en_velocity}
\title{Compute Euclidean norm of sums of velocity vectors (x,y)}
\usage{
en_velocity(vel_x, vel_y)
}
\arguments{
\item{vel_x}{Data frame of x-axis velocity generated by velocity_x().}
\item{vel_y}{Data frame of y-axis velocity generated by velocity_y().}
}
\description{
Compute Euclidean norm of sums of velocity vectors (x,y) for a tracked video using the x-axis and y-axis data generated using velocity_x() and velocity_y(). This is done using the following formula: sqrt(velocityx(t+1) - velocityx(t))^2 + (velocityy(t+1) - velocityy(t))^2)
}
\examples{
en_velocity(x, y)
}
